cambios realizados con validaciones en movieControlers

+++ Controllers/MoviesController.cs
@@ -132,26 +132,6 @@ namespace AppMovie.Controllers
             return View(movie);
         }
 
-        // GET: Movies/Delete/5
-        public async Task<IActionResult> Delete(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var movie = await _context.Movie
-                .Include(m => m.Gender)
-                .Include(m => m.Producer)
-                .Include(m => m.Section)
-                .FirstOrDefaultAsync(m => m.MovieID == id);
-            if (movie == null)
-            {
-                return NotFound();
-            }
-
-            return View(movie);
-        }
 
         // POST: Movies/Delete/5
         // [HttpPost, ActionName("Delete")]
@@ -160,8 +140,18 @@ namespace AppMovie.Controllers
         {
             var movie = await _context.Movie.FindAsync(id);
             if(movie != null){
-            _context.Movie.Remove(movie);
-            await _context.SaveChangesAsync();
+                var movieAlquilada = (from a in _context.RentalDetail where a.MovieID == id select a).Count();
+                if(movieAlquilada == 0)
+                {
+                    _context.Movie.Remove(movie);
+                    await _context.SaveChangesAsync();
+                }else
+                {
+                    movie.IsDeleted = true;
+                    movie.MovieName = "Eliminada";
+                    _context.Update(movie);
+                    await _context.SaveChangesAsync();
+                }
             }
 
             return RedirectToAction(nameof(Index));
